requirements.txt
--extra-index-url https://download.pytorch.org/whl/nightly/cu126
torch
torchvision
torchaudio
supervision
trackers[deepsort,cu126]
git+https://github.com/huggingface/transformers.git
opencv-python
Pillow
gspread
google-auth
pyyaml
fastapi
uvicorn
pydantic
mediapipe
filterpy
lap
scikit-image
huggingface-hub
alive-progress
python-dotenv